+++ test/do.js
@@ -1,3 +1,0 @@
-window.returnTitle = function() {
-  return document.title;
-};