+++ test/simple.js
@@ -1,16 +1,27 @@
-// const page = require('webpage').create();
+describe('preflight checks', function() {
+    it('tests are running', function() {
+        expect(1 + 1).toBe(2);
+    });
 
-// page.open('http://localhost:9090/test', function (status) {
-//      if (status !== 'success') {
-//         console.log('Unable to access network');
-//     } else {
-//         console.log( 'Opened ok' );
-//         page.includeJs('./../dist/lightning.js', function() {
-//             console.log('Lightning included');
-//             console.log(Lightning);
-//             // var rng = window.Lightning.Maths.rngInt(0, 20);
-//             // console.log(rng);
-//             phantom.exit();
-//         });
-//     }
-// });
+    it('window exists', function() {
+        expect(window).toBeDefined(); 
+    });
+
+    it('document exists', function() {
+        expect(document).toBeDefined(); 
+    });
+
+    it('lightning exist', function() {
+        expect(window.Lightning).toBeDefined(); 
+    });
+});
+
+describe('util functions', function() {
+    var Lightning = window.Lightning;
+    
+    // it('convert hex to rgb', function() {
+    //     var hex = 0xffffff;
+        
+    //     expect(Lightning.Maths.hextoRGB(0xffffff)).toEqual([255,255,255]);
+    // });
+});